refactor: adjust directory structure and imports, add import alias
