+++ packages/core/src/index.ts
@@ -1,11 +1,4 @@
-export * from './actor'
-export * from './feed'
-export * from './list'
-export * from './paginate'
-export * from './post'
-export * from './preference'
-export * from './starterPack'
+export * from './bsky'
 export * from './tsky'
-export * from './video'
 
 export { TSky as default } from './tsky'
