+++ packages/core/src/tsky/paginate.ts
